refactor: fix tests, upgrade ant
