+++ package.json
@@ -7,7 +7,7 @@
     "@ant-design/icons": "^5.2.6",
     "ag-grid-community": "^31.0.3",
     "ag-grid-react": "^31.0.1",
-    "antd": "^5.14.1",
+    "antd": "^5.15.1",
     "btoa": "^1.2.1",
     "fast-sort": "^3.4.0",
     "html-to-image": "^1.11.11",
